#220 - Polishing.
Formatting.
Fix warnings.
JavaDoc.
Adjusted nullability annotations.

Original pull request: #287.

+++ src/main/java/org/springframework/data/r2dbc/core/DefaultDatabaseClient.java
@@ -309,9 +309,9 @@ class DefaultDatabaseClient implements DatabaseClient, ConnectionAccessor {
 		byIndex.forEach((i, o) -> {
 
 			if (o.getValue() != null) {
-				statement.bind(i.intValue(), o.getValue());
+				statement.bind(i, o.getValue());
 			} else {
-				statement.bindNull(i.intValue(), o.getType());
+				statement.bindNull(i, o.getType());
 			}
 		});
 	}
